remove silly license header from file
